requirements.txt
Flask==2.0.1
SpeechRecognition==3.8.1
pytesseract==0.3.8
Pillow==9.0.0
pydub==0.25.1
Werkzeug==2.0.1
gunicorn==20.1.0
python-dotenv==0.19.1
Flask-Cors==3.0.10